[BUGFIX]Squash easy warnings to reduce default warning level noise when compiling

+++ client/gui/gui_boot.cpp
@@ -313,7 +313,7 @@ class BootWindow : public Fl_Window
 		BootWindow* boot = static_cast<BootWindow*>(data);
 
 		const int val = boot->m_PWADOrderBrowser->value() - 1;
-		if (val <= 0 || val >= boot->m_selectedPWADs.size())
+		if (val <= 0 || val >= static_cast<int>(boot->m_selectedPWADs.size()))
 			return;
 
 		std::iter_swap(boot->m_selectedPWADs.begin() + val,
@@ -327,7 +327,7 @@ class BootWindow : public Fl_Window
 		BootWindow* boot = static_cast<BootWindow*>(data);
 
 		const int val = boot->m_PWADOrderBrowser->value() - 1;
-		if (val < 0 || val >= boot->m_selectedPWADs.size() - 1)
+		if (val < 0 || val >= static_cast<int>(boot->m_selectedPWADs.size() - 1))
 			return;
 
 		std::iter_swap(boot->m_selectedPWADs.begin() + val,
@@ -395,7 +395,7 @@ class BootWindow : public Fl_Window
 		BootWindow* boot = static_cast<BootWindow*>(data);
 
 		const int val = boot->m_WADDirList->value() - 1;
-		if (val <= 0 || val >= boot->m_WADDirs.size())
+		if (val <= 0 || val >= static_cast<int>(boot->m_WADDirs.size()))
 			return;
 
 		std::iter_swap(boot->m_WADDirs.begin() + val, boot->m_WADDirs.begin() + val - 1);
@@ -409,7 +409,7 @@ class BootWindow : public Fl_Window
 		BootWindow* boot = static_cast<BootWindow*>(data);
 
 		const int val = boot->m_WADDirList->value() - 1;
-		if (val < 0 || val >= boot->m_WADDirs.size() - 1)
+		if (val < 0 || val >= static_cast<int>(boot->m_WADDirs.size() - 1))
 			return;
 
 		std::iter_swap(boot->m_WADDirs.begin() + val, boot->m_WADDirs.begin() + val + 1);
@@ -426,7 +426,7 @@ class BootWindow : public Fl_Window
 		BootWindow* boot = static_cast<BootWindow*>(data);
 
 		const int val = boot->m_WADDirList->value() - 1;
-		if (val < 0 || val >= boot->m_WADDirs.size())
+		if (val < 0 || val >= static_cast<int>(boot->m_WADDirs.size()))
 			return;
 
 		boot->m_WADDirs.erase(boot->m_WADDirs.begin() + val);
